requirements.txt
python-telegram-bot==20.3
requests==2.31.0
numpy==1.26.4
pandas==2.1.4
PyYAML==6.0.1